fix type of sa_index_t as per funny falcon

+++ include/ruby/st.h
@@ -151,7 +151,7 @@ st_index_t st_hash_start(st_index_t h);
 #pragma GCC visibility pop
 #endif
 
-typedef uint64_t sa_index_t;
+typedef uint32_t sa_index_t;
 #define SA_STOP     ST_STOP
 #define SA_CONTINUE ST_CONTINUE
